debug: Debug-missing-wake
Under certain scenarios async-io can miss wake's. I've seen this in
libp2p when running libp2p-relay on a small Digital Ocean VM with
1 CPU 1GB ram with RUST_LOG=trace and a debug build. And then run
libp2p-lookup. Most of the times it fails.

I'm going to try to create a test that shows the issue and then try to
devise a solution.