fix: Change setting of ticks
Try NOT using the current Reactor.ticker, just use the current tick.
The reason is we don't know if the current Reactor.ticker actually has
been used and in this bug that is definitely the case.

The other option is to use None, but this is "true" so seems reasonable.